requirements.txt
python-dotenv
networkx
psycopg2
docling
requests
fastapi
uvicorn
torch
future
transformers
accelerate
sentence-transformers
graspologic
nltk
kuzu
pydantic
faiss-cpu
pip-licenses
